Always allow menu access
Custom click action now only is for the primary button.

This means the user can use for example "select area" as the primary
button action and access the extension menu with the secondary button.

Previously, the user could not access the settings menu from the
indicator when a custom action was set.